Moved print functions to test-core
Also renamed *_print to print_*
