+++ mft.c
@@ -189,6 +189,7 @@ mft_parse_econtent(const unsigned char *d, size_t dsz, struct parse *p, int forc
 	struct tm			tm;
 	int			 i, rc = -1;
 	time_t			 this, next, now = time(NULL);
+	char 			 buf[64];
 	char			 caThis[64], caNow[64], caNext[64];
 
 	if ((seq = d2i_ASN1_SEQUENCE_ANY(NULL, &d, dsz)) == NULL) {
@@ -268,7 +269,7 @@ mft_parse_econtent(const unsigned char *d, size_t dsz, struct parse *p, int forc
 
 	memcpy(&p->res->thisUpdate, &this, sizeof (time_t));
 	memcpy(&p->res->nextUpdate, &next, sizeof (time_t));
-/*
+
 	if (this >= next) {
 		warnx("%s: bad update interval [%s] >= [%s]", p->fn, caThis, caNext);
 		goto out;
@@ -286,7 +287,7 @@ mft_parse_econtent(const unsigned char *d, size_t dsz, struct parse *p, int forc
 		}
 		if (verbose > 0)
 			warnx("%s: stale: expired %s (ignoring)", p->fn, buf);
-	}*/
+	}
 
 	/* File list algorithm. */
