+++ test-crl.c
@@ -16,10 +16,7 @@
  */
 #include "config.h"
 
-#include <assert.h>
-#include <err.h>
 #include <stdio.h>
-#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -31,69 +28,6 @@
 #include "test-core.h"
 
 int	 verbose;
-
-// http://www.geo-complex.com/shares/soft/unix/CentOS/OpenVPN/openssl-1.1.0c/crypto/x509/x_crl.c
-static void
-crl_print(const X509_CRL *p)
-{
-	int i, numRevoked;
-	char caRevocationDate[64];
-	char caLast[64], caNext[64], caNow[64];
-	char *issuerName;
-	time_t now;
-	struct tm tm;
-	STACK_OF(X509_REVOKED) *revoked;
-
-	assert(p != NULL);
-
-    now = time(NULL);
-	strftime(caNow, sizeof(caNow)-1, "%Y-%m-%d %H:%M:%S GMT", gmtime(&now));
-
-	revoked = X509_CRL_get_REVOKED(p);
-
-	tm = asn1Time2Time(p->crl->lastUpdate);
-	strftime(caLast, sizeof(caLast)-1, "%Y-%m-%d %H:%M:%S GMT", &tm);
-
-	tm = asn1Time2Time(p->crl->nextUpdate);
-	strftime(caNext, sizeof(caNext)-1, "%Y-%m-%d %H:%M:%S GMT", &tm);
-
-	printf("%*.*s: %s\n", TAB, TAB, "Now", caNow);
-	print_sep_line("Certificate Revocation List", 110);
-	printf("%*.*s: %ld\n", TAB, TAB, "Version", ASN1_INTEGER_get(p->crl->version) + 1);
-	printf("%*.*s: %ld\n", TAB, TAB, "CRL Number", ASN1_INTEGER_get(p->crl_number));
-
-	issuerName = X509_NAME_oneline(p->crl->issuer, NULL, 0);
-	if (issuerName != NULL) {
-		printf("%*.*s: %s\n", TAB, TAB, "Issuer", issuerName);
-		OPENSSL_free(issuerName);
-	}
-
-	printf("%*.*s: %s\n", TAB, TAB, "Last Update", caLast);
-	printf("%*.*s: %s\n", TAB, TAB, "Next Update", caNext);
-
-	numRevoked = sk_X509_REVOKED_num(revoked);
-	if (numRevoked > 0) {
-		print_sep_line("Revoked Certificates", 110);
-		for (i = 0; i < numRevoked; i++) {
-			X509_REVOKED *rev = sk_X509_REVOKED_value(revoked, i);
-			if (rev != NULL) {
-				BIGNUM *bnSrl = ASN1_INTEGER_to_BN(rev->serialNumber, NULL);
-				if (bnSrl != NULL) {
-					char *lpcSrl = BN_bn2hex(bnSrl);
-					if (lpcSrl != NULL) {
-						printf("%*.*s: %s\n", TAB, TAB, "Serial Number", lpcSrl);
-						OPENSSL_free(lpcSrl);
-					}
-					BN_free(bnSrl);
-				}
-			}
-			tm = asn1Time2Time(rev->revocationDate);
-			strftime(caRevocationDate, sizeof(caRevocationDate)-1, "%Y-%m-%d %H:%M:%S GMT", &tm);
-			printf("%*.*s:    %s\n", TAB, TAB, "Revokation Date", caRevocationDate);
-		}
-	}
-}
-
 
 int
 main(int argc, char *argv[])
@@ -121,7 +55,7 @@ main(int argc, char *argv[])
 		if ((p = crl_parse(argv[i], NULL)) == NULL)
 			break;
 		if (verbose)
-			crl_print(p);
+			print_crl(p);
 		X509_CRL_free(p);
 	}
